+++ src/RuStoreClient.php
@@ -6,9 +6,9 @@ namespace NotificationChannels\RuStore;
 
 use GuzzleHttp\Promise\PromiseInterface;
 use Illuminate\Http\Client\Response;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Http;
 use NotificationChannels\RuStore\Exceptions\RuStorePushException;
-use NotificationChannels\RuStore\Reports\RuStoreReport;
 use NotificationChannels\RuStore\Reports\RuStoreSingleReport;
 use Throwable;
 
@@ -17,6 +17,7 @@ class RuStoreClient
     // @todo задействовать проверку максимального объема сообщения
     public const MAX_PAYLOAD_LENGTH = 4096;
 
+    // @todo вынести в настройки?
     private const URL_FORMAT = 'https://vkpns.rustore.ru/v1/projects/%s/messages:send';
     private readonly string $url;
     private readonly string $bearer_token;
@@ -31,17 +32,14 @@ class RuStoreClient
      * Отправка уведомлений на все устройства пользователя
      *
      * @param RuStoreMessage $message
-     * @param array $tokens
-     * @return RuStoreReport
+     * @param array<int, string> $tokens
+     * @return Collection<int, RuStoreSingleReport>
      */
-    public function send(RuStoreMessage $message, array $tokens): RuStoreReport
+    public function send(RuStoreMessage $message, array $tokens): Collection
     {
-        $report = RuStoreReport::init($tokens, $message);
-        $report->all()->each(function (?RuStoreSingleReport $_, string $token) use ($report, $message): void {
-            $report->addReport($token, $this->sendSingle($message, $token));
-        });
-
-        return $report;
+        // @todo проверить тип $token!
+        // return collect(array_combine($tokens, $tokens))
+        return collect($tokens)->map(fn(string $token): RuStoreSingleReport => $this->sendSingle($message, $token));
     }
 
     /**
@@ -59,11 +57,13 @@ class RuStoreClient
             $response = $request->send('POST', $this->url);
 
         } catch (Throwable $exception) {
-            return RuStoreSingleReport::failure($exception);
+            return RuStoreSingleReport::failure($token, $exception);
         }
 
+        // dump($response->successful());
+
         return $response->successful()
-            ? RuStoreSingleReport::success($response)
-            : RuStoreSingleReport::failure(RuStorePushException::fromResponse($response), $response);
+            ? RuStoreSingleReport::success($token, $response)
+            : RuStoreSingleReport::failure($token, RuStorePushException::fromResponse($response), $response);
     }
 }